requirements.txt
Django==1.6.6
Markdown==2.5.1
Pillow==2.5.3
PyJWT==0.2.1
South==1.0
Werkzeug==0.9.6
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.4.0
django-configurations==0.8
django-cors-headers==0.13
django-countries==2.1.2
django-debug-toolbar==1.2.1
django-dotenv==1.2
django-extensions==1.3.11
django-gravatar2==1.1.4
django-rest-swagger==0.1.14
django-reversion==1.8.2
django-secure==1.0
django-taggit==0.12.1
djangorestframework==2.3.14
djangorestframework-jwt==1.0.1
djrill==1.1.1
gunicorn==19.1.1
jsonfield==1.0.0
psycopg2==2.5.4
requests==2.3.0
six==1.7.3
sqlparse==0.1.11
static3==0.5.1
wsgiref==0.1.2
